WIP zone file format document
Still largely incomplete. I plan I detailing all supported properties for the general object and MP object section + nav_node and cover_node objects in the SP section. The rest of the SP section and the runtime section won't be filled out until the SP map support is added to the Nanoforge rewrite.